final part of unit 0x02 added

+++ 0x02/snippets/c_control_flow/self_a_control_flow.py
@@ -123,7 +123,7 @@ def using_exceptions():
 
     print(" 6| division by 0...", end='')
     try:
-        n = 1/0
+        n = int(1/0)
         print(f" {n=}")
     except Exception as e:                                      # pylint: disable=broad-exception-caught
         print(f" => unknown error: {e}")                        # optional: e.with_traceback()
